gpu: ipu-v3: compress map definitions
The i.MX6 IPU DC supports 30 color maps (consisting of 3 pointers per
entry) but has only 24 slots (which are the target of these pointers).
A "color map" is an index used in IPU microcode to transform three color
components.  This transformation is expressed by an offset + mask pair
which is stored in the slots mentioned above.

Previous code used a very simple method to calculate the slot number:

| int ptr = map * 3 + byte_num

This means, that only 24/3 = 8 maps (from the possible 30 ones) can be
used.  This suffices for simple cases, but when working e.g. with
asynchronous displays, more transformations are required.

Because the offset + mask pairs are repeating in many cases and only
their combination differs, they can be stored more efficiently.

E.g. when there are two color maps

 A = [ ( 7, 0xff), (15, 0xff), (23, 0xff) ]      /* RGB24 */
 B = [ (23, 0xff), (15, 0xff), ( 7, 0xff) ]      /* BGR24 */

The previous code allocated 6 slots for them

 S[0] = ( 7, 0xff)
 S[1] = (15, 0xff)
 S[2] = (23, 0xff)
 -
 S[3] = (23, 0xff)
 S[4] = (15, 0xff)
 S[5] = ( 7, 0xff)

and defined the maps as

 M[A] = (0, 1, 2)
 M[B] = (3, 4, 5)

With this patch, slots are reused so that only S[0], S[1] and S[2] are
allocated and the mappings are now

 M[A] = (0, 1, 2)
 M[B] = (2, 1, 0)

Code size is increased by

$ ./scripts/bloat-o-meter drivers/gpu/ipu-v3/ipu-dc.o{__,}
add/remove: 1/0 grow/shrink: 3/1 up/down: 912/-308 (604)
function                                     old     new   delta
ipu_dc_register_mappings                       -     812    +812
descriptor                                    48      96     +48
__func__                                      29      73     +44
ipu_dc_init_sync                            1112    1120      +8
ipu_dc_init                                  844     536    -308

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>