gpu: ipu-v3: use a more fitting name for 'display' attribute
Avoid confusion in the driver by using 'display' instead of 'di' as the
name of the attribute which specifies the display.  'di' in IPU context
means the display interface and there is only a loose coupling of display
and di.

E.g. IPU supports 4 displays but has only 2 DI units.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>