gpu: ipu-v3: allow variable expansion in microcode
When writing more complicated microcode (--> more instructions), it
might be desired to store it into 'static const' memory (.rodata or
.text).  Variable parts (e.g. color map idx, addresses or wait cycles)
can prevent this so that microcode must be created on stack at runtime.

This patch adds a simple linker to the microcode generator.  The
required information are encoded in the upper, unused 22 bits of the
64 bit microcode.

For symbol substitution, these 22 bits contain a pointer to a continuous
area in the microcode (--> position and width).  The actual static value
of this area will be used as an index of a dynamic symbol table.  The
area will be zeroed and replaced by the value in this table.

For jump instructions, the address part in the static microcode has now
the meaning of a relative offset to the address of the jump instruction.

Code size increases by this commit but most of this space is static data
in readonly memory.

add/remove: 6/1 grow/shrink: 2/0 up/down: 760/-364 (396)
function                                     old     new   delta
ipu_dc_init_sync                             584    1076    +492
events_non_interlaced                          -     108    +108
events_interlaced                              -     108    +108
microcode_new_data                             -      16     +16
microcode_eol                                  -      16     +16
microcode_nl                                   -       8      +8
microcode_interlaced                           -       8      +8
ipu_dc_register_mappings                     812     816      +4
ipu_dc_write_evt_microcodes.constprop        364       -    -364

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>